requirements.txt
streamlit
streamlit_option_menu
langchain-core 
langchain-groq 
python-dotenv